+++ src/jep/pyjmethod.h
@@ -68,7 +68,7 @@ int pyjmethod_check(PyObject *obj);
 /*
  * Test if a method has a specified name and number of parameters.
  */
-int pyjmethod_check_simple_compat(PyJmethod_Object*, JNIEnv*, PyObject*, int);
+int pyjmethod_check_simple_compat(PyJmethod_Object*, JNIEnv*, PyObject*, Py_ssize_t);
 
 /*
  * Check if a method is compatible with the types of a tuple of arguments. It is
